problematy/requirements.txt
asgiref==3.7.2
Django==5.0.3
django-model-utils==4.5.0
django-widget-tweaks==1.5.0
pillow==10.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.4.4
gunicorn==20.1.0
google-cloud-storage==2.17.0